[Bronze II] Title: Hashing, Time: 100 ms, Memory: 14216 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -5,10 +5,10 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         
-        int n = Integer.parseInt(br.readLine());
+        long n = Long.parseLong(br.readLine());
         String s = br.readLine();
-        int sum = 0;
-        int i = 1;
+        long sum = 0;
+        long i = 1;
         
         for (char c : s.toCharArray()) {
             sum += (c - 'a' + 1) * i;
